Migrates ?aoc to /aoc
