+++ test/commands/slash/AdventofcodeCommandTest.ts
@@ -1,9 +1,7 @@
 import { createSandbox, SinonSandbox } from "sinon";
 import { expect } from "chai";
-import { Message } from "discord.js";
 import { BaseMocks } from "@lambocreeper/mock-discord.js";
-
-import AdventofcodeCommand from "../../../src/commands/legacy/AdventofcodeCommand";
+import AdventofcodeCommand from "../../../src/commands/slash/AdventofcodeCommand";
 import AdventOfCodeService from "../../../src/services/AdventOfCodeService";
 import { EMBED_COLOURS, ADVENT_OF_CODE_INVITE, ADVENT_OF_CODE_LEADERBOARD, ADVENT_OF_CODE_RESULTS_PER_PAGE } from "../../../src/config.json";
 import { AOCLeaderBoard } from "../../../src/interfaces/AdventOfCode";
@@ -31,98 +29,79 @@ const AOCMockData: AOCLeaderBoard = {
 };
 
 describe("Adventofcode Command", () => {
-	describe("constructor", () => {
-		it("creates a command called adventofcode", () => {
-			const command = new AdventofcodeCommand();
-
-			expect(command.getName()).to.equal("adventofcode");
-			expect(command.getAliases().includes("aoc")).to.be.true;
-		});
-
-		it("creates a command with correct description", () => {
-			const command = new AdventofcodeCommand();
-
-			expect(command.getDescription()).to.equal("Shows the current leaderboard for adventofcode.");
-		});
-	});
-
-	describe("run()", () => {
+	describe("onInteract()", () => {
 		let sandbox: SinonSandbox;
-		let message: Message;
 		let command: AdventofcodeCommand;
+		let interaction: any;
+		let replyStub: sinon.SinonStub<any[], any>;
 		let AOC: AdventOfCodeService;
 
 		beforeEach(() => {
 			sandbox = createSandbox();
 			command = new AdventofcodeCommand();
-			message = BaseMocks.getMessage();
+			replyStub = sandbox.stub().resolves();
+			interaction = {
+				reply: replyStub,
+				user: BaseMocks.getGuildMember()
+			};
 			AOC = AdventOfCodeService.getInstance();
 		});
 
 		it("sends a message to the channel", async () => {
-			const messageMock = sandbox.stub(message.channel, "send");
-
 			sandbox.stub(AOC, "getLeaderBoard").resolves(AOCMockData);
 
-			await command.run(message, []);
+			await command.onInteract(undefined, interaction);
 
-			expect(messageMock.calledOnce).to.be.true;
+			expect(replyStub.calledOnce).to.be.true;
 		});
 
 		it("sends an error message when the year is too far into the feature", async () => {
-			const messageMock = sandbox.stub(message.channel, "send");
-
 			sandbox.stub(AOC, "getLeaderBoard").throws();
 			sandbox.stub(command, "getYear").returns(2019);
 
-			await command.run(message, ["2021"]);
-			const embed = messageMock.getCall(0).firstArg.embeds[0];
+			await command.onInteract("2021", interaction);
+			const embed = replyStub.getCall(0).firstArg.embeds[0];
 
-			expect(messageMock.calledOnce).to.be.true;
+			expect(replyStub.calledOnce).to.be.true;
 			expect(embed.title).to.equal("Error");
 			expect(embed.description).to.equal("Year requested not available.\nPlease query a year between 2015 and 2019");
 			expect(embed.hexColor).to.equal(EMBED_COLOURS.ERROR.toLowerCase());
 		});
 
 		it("sends an error message when the year is too far in the past", async () => {
-			const messageMock = sandbox.stub(message.channel, "send");
-
 			sandbox.stub(AOC, "getLeaderBoard").throws();
 			sandbox.stub(command, "getYear").returns(2018);
 
-			await command.run(message, ["2000"]);
-			const embed = messageMock.getCall(0).firstArg.embeds[0];
+			await command.onInteract("2000", interaction);
+			const embed = replyStub.getCall(0).firstArg.embeds[0];
 
-			expect(messageMock.calledOnce).to.be.true;
+			expect(replyStub.calledOnce).to.be.true;
 			expect(embed.title).to.equal("Error");
 			expect(embed.description).to.equal("Year requested not available.\nPlease query a year between 2015 and 2018");
 			expect(embed.hexColor).to.equal(EMBED_COLOURS.ERROR.toLowerCase());
 		});
 
 		it("should query the year 2018 from the AOC Service", async () => {
-			const messageMock = sandbox.stub(message.channel, "send");
 			const serviceMock = sandbox.stub(AOC, "getLeaderBoard").resolves({ members: {}, event: "2018", owner_id: "12345" });
 
 			sandbox.stub(command, "getYear").returns(2020);
 
-			await command.run(message, ["2018"]);
+			await command.onInteract("2018", interaction);
 
 			expect(serviceMock.calledOnce).to.be.true;
 			expect(serviceMock.getCall(0).args[1]).to.equal(2018);
-			expect(messageMock.calledOnce).to.be.true;
+			expect(replyStub.calledOnce).to.be.true;
 		});
 
 		it("sends a message with the current score", async () => {
-			const messageMock = sandbox.stub(message.channel, "send");
-
 			sandbox.stub(AOC, "getLeaderBoard").resolves(AOCMockData);
 			sandbox.stub(command, "getYear").returns(2019);
 
-			await command.run(message, []);
+			await command.onInteract(undefined, interaction);
 
-			const embed = messageMock.getCall(0).firstArg.embeds[0];
+			const embed = replyStub.getCall(0).firstArg.embeds[0];
 
-			expect(messageMock.calledOnce).to.be.true;
+			expect(replyStub.calledOnce).to.be.true;
 			expect(embed.title).to.equal("Advent Of Code");
 			expect(embed.description).to.equal(`Leaderboard ID: \`${ADVENT_OF_CODE_INVITE}\`\n\n[View Leaderboard](https://adventofcode.com/2019/leaderboard/private/view/${ADVENT_OF_CODE_LEADERBOARD})`);
 			expect(embed.fields[0].name).to.equal(`Top ${ADVENT_OF_CODE_RESULTS_PER_PAGE}`);
@@ -131,31 +110,27 @@ describe("Adventofcode Command", () => {
 		});
 
 		it("gives an error when the wrong acces token/id is provided", async () => {
-			const messageMock = sandbox.stub(message.channel, "send");
-
 			sandbox.stub(AOC, "getLeaderBoard").throws();
 
-			await command.run(message, []);
+			await command.onInteract(undefined, interaction);
 
-			const embed = messageMock.getCall(0).firstArg.embeds[0];
+			const embed = replyStub.getCall(0).firstArg.embeds[0];
 
-			expect(messageMock.calledOnce).to.be.true;
+			expect(replyStub.calledOnce).to.be.true;
 			expect(embed.title).to.equal("Error");
 			expect(embed.description).to.equal("Could not get the leaderboard for Advent Of Code.");
 			expect(embed.hexColor).to.equal(EMBED_COLOURS.ERROR.toLowerCase());
 		});
 
 		it("gives back one user when giving a username argument", async () => {
-			const messageMock = sandbox.stub(message.channel, "send");
-
 			sandbox.stub(AOC, "getLeaderBoard").resolves(AOCMockData);
 			sandbox.stub(command, "getYear").returns(2021);
 
-			await command.run(message, ["Lambo"]);
+			await command.onInteract("Lambo", interaction);
 
-			const embed = messageMock.getCall(0).firstArg.embeds[0];
+			const embed = replyStub.getCall(0).firstArg.embeds[0];
 
-			expect(messageMock.calledOnce).to.be.true;
+			expect(replyStub.calledOnce).to.be.true;
 			expect(embed.title).to.equal("Advent Of Code");
 			expect(embed.description).to.equal(`Leaderboard ID: \`${ADVENT_OF_CODE_INVITE}\`\n\n[View Leaderboard](https://adventofcode.com/2021/leaderboard/private/view/${ADVENT_OF_CODE_LEADERBOARD})`);
 			expect(embed.fields[0].name).to.equal("Scores of Lambo");
@@ -170,31 +145,28 @@ describe("Adventofcode Command", () => {
 		});
 
 		it("should give an error when the user doesn't exist", async () => {
-			const messageMock = sandbox.stub(message.channel, "send");
-
 			sandbox.stub(AOC, "getLeaderBoard").resolves(AOCMockData);
 
-			await command.run(message, ["Bob"]);
+			await command.onInteract("Bob", interaction);
 
-			const embed = messageMock.getCall(0).firstArg.embeds[0];
+			const embed = replyStub.getCall(0).firstArg.embeds[0];
 
-			expect(messageMock.calledOnce).to.be.true;
+			expect(replyStub.calledOnce).to.be.true;
 			expect(embed.title).to.equal("Error");
 			expect(embed.description).to.equal("Could not get the user requested\nPlease make sure you typed the name correctly");
 			expect(embed.hexColor).to.equal(EMBED_COLOURS.ERROR.toLowerCase());
 		});
 
 		it("requesting a different year then current", async () => {
-			const messageMock = sandbox.stub(message.channel, "send");
 			const APIMock = sandbox.stub(AOC, "getLeaderBoard").resolves(AOCMockData);
 
 			sandbox.stub(command, "getYear").returns(2020);
 
-			await command.run(message, ["2019"]);
+			await command.onInteract("2019", interaction);
 
-			const embed = messageMock.getCall(0).firstArg.embeds[0];
+			const embed = replyStub.getCall(0).firstArg.embeds[0];
 
-			expect(messageMock.calledOnce).to.be.true;
+			expect(replyStub.calledOnce).to.be.true;
 			expect(APIMock.getCall(0).args[1]).to.equal(2019);
 			expect(embed.title).to.equal("Advent Of Code");
 			expect(embed.description).to.equal(`Leaderboard ID: \`${ADVENT_OF_CODE_INVITE}\`\n\n[View Leaderboard](https://adventofcode.com/2019/leaderboard/private/view/${ADVENT_OF_CODE_LEADERBOARD})`);
